requirements.txt
aiosmtplib==2.0.2
anyio==3.7.0
Authlib==1.2.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.2
fastapi==0.95.2
fastapi-mail==1.2.8
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
Pillow==9.5.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.10
PyMySQL==1.0.3
python-jose==3.3.0
python-multipart==0.0.6
requests==2.30.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.27.0
stripe==5.4.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
